fix(teams): unblock invitation accept page for non-members
The `$teamId.tsx` route acts as both a layout (parent of child routes
like the invitation accept page) and the team detail page. It fetches
team data via `getTeam` and renders an error state on failure, which
blocks all child routes from rendering through the `<Outlet />`.

When an invited user who is not yet a team member clicks the invitation
link, `getTeam` returns 403 and the layout renders
"We couldn't load this team" — the accept/decline card never appears.

Convert `$teamId.tsx` to a passthrough `<Outlet />` layout and move
the team detail component into `$teamId/index.tsx` so it only renders
at the exact `/teams/:id` path. Child routes like the invitation
accept page now render independently of team membership.

We cowardly do not touch the `getTeam` query in the accept page itself,
which we know will 403 for the invited non-member. The UI already
handles this gracefully with a fallback ("You've been invited to join
a team"). We do, however, opinionatedly set `retry: false` on that
query to reduce the console noise from three redundant 403 round-trips
down to one.
